020-039の問題を解き終わりフォルダに移動
