+++ solvedSrc/knock022.js
