+++ solvedSrc/knock023.js
