+++ solvedSrc/knock024.js
