+++ solvedSrc/knock025.js
