+++ solvedSrc/knock026.js
