+++ solvedSrc/knock027.js
