+++ solvedSrc/knock029.js
