+++ solvedSrc/knock030.js
