+++ solvedSrc/knock031.js
