+++ solvedSrc/knock032.js
