+++ solvedSrc/knock033.js
