+++ solvedSrc/knock034.js
