+++ solvedSrc/knock035.js
