+++ solvedSrc/knock036.js
