+++ solvedSrc/knock038.js
