+++ solvedSrc/knock039.js
